+++ C++/convolution/dual_convol.h
@@ -8,11 +8,11 @@
 class DualConvol : public Convol
 {
   public:
-    void apply_filter(double* out, double* padded_img, Filter filter, int width, int height, int padding);
+    void apply_filter(uint8_t* out, uint8_t* padded_img, Filter filter, int width, int height, int padding);
     void whatIsMyName();
 
   private:
-    void convol(double* out, double* img, Filter filter, int x, int y, int stride);
+    void convol(uint8_t* out, uint8_t* img, Filter filter, int x, int y, int stride);
 };
 
 #endif
